+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ChooseActionScreen.kt
@@ -55,8 +55,7 @@ fun HandleActionBottomSheets(delegate: CreateActionDelegate) {
     ChangeFlashlightStrengthActionBottomSheet(delegate)
     HttpRequestBottomSheet(delegate)
     SmsActionBottomSheet(delegate)
-    VolumeUpActionBottomSheet(delegate)
-    VolumeDownActionBottomSheet(delegate)
+    VolumeActionBottomSheet(delegate)
 }
 
 @Composable